nasm: add %included assembly files to nasm_assemble action
This is to simplify siso config and utilize GN's input handling
functionality.

Currently, we have
https://source.chromium.org/chromium/chromium/src/+/main:build/config/siso/nasm_scandeps.star
to make nasm actions with missing inputs run remotely when we use
siso.

But the config shouldn't be necessary if inputs are specified
correctly even if they are indirect inputs in nasm actions.
To remove the unnecessary config, this CL specifies included assembly
files as inputs in BUILD.gn.

Bug: 358521078
Change-Id: I3191cfb2c50cb5b27776d0c26a52a2ad70db798b